First try at syncing movementdata
If you are the authority, send a command that sends an rpc that tells clients to change their movement data.
Otherwise, lerp towards estimatedposition
Fixed issue where PMove's PCon wouldn't be assigned on players, because connection variable wasn't a syncvar.
The syncing is very bad, so the lerping is extreme, but both players are spawned correctly and are able to move.