For loop solution
We implemented the for-loop method for doing movement prediction. It works a little bit better, but is still largely broken.
On the host, the other player appears to be in two different places at once, where one of them moves correctly, and the other falls down way below screen and then appears again higher than the regular bounce reaches. This movement is regular.
The client sees the other player in a more strange way, where the balls jumps between many positions, and reaches higher than it normally would.