Static rigidbody solution
We made a working solution by setting the rigidbody to static when !hasAuthority. The makes the program run smoothly, but it works poorly on higher rtt, since the transform either always has to lerp or teleport, instead of running the physics in parallel with the estimation. Apparently, running the physics custom on the estimation while also running the rigidbody caused issues for the rigidbody.