Mirror movement now works!
We forgot to adjust movement by deltatime during physics. This means the estimated position is moving too fast, and only looks reasonable because the rpc keeps teleporting it back. Sometimes, the rpc is not called in time to move it back, and the rigidbody moves over too much.
Still missing movement validation check, player-player collisions and shooting