Fixed movement syncing
When moving forwards, the other player moved strangely, because the other player didn't get updates on what the angle was (it was updated in GetInput()), and so always moved to the right. Now fixed. Still a little bit wonky when changing directions and hitting walls, but with a smooth lerp it's not too bad.